Robustify completion match scoring for incomplete/implicit patterns
completion-pcm--hilit-commonality, which propertizes and scores a
previously confirmed match, expects PATTERN to match the strings `of
COMPLETIONS entirely.  In this patch, we take care that the main goal
of this function, isn't affected by some optimzations that transform
PATTERN beforehand.

* lisp/minibuffer.el (completion-pcm--hilit-commonality):
Initialize start to 0 and end to the length of the string.